third_party/blink/web_tests/http/tests/devtools/sources/debugger-ui/function-details-expected.txt

Tests that Debugger.getFunctionDetails command returns correct location. Bug 71808


Running: testGetFirstLineFunctionDetails
firstLineFunction type = function
Function details: 
lineNumber: 12
columnNumber: 34
scriptId is valid: true
functionName: firstLineFunction

Running: testGetNonFirstLineFunctionDetails
notFirstLineFunction type = function
Function details: 
lineNumber: 17
columnNumber: 37
scriptId is valid: true
functionName: notFirstLineFunction

Running: testGetDetailsOfFunctionWithInferredName
obj.m type = function
Function details: 
lineNumber: 23
columnNumber: 23
scriptId is valid: true
functionName: m

Running: testGetDetailsOfFunctionWithDisplayName
functionWithDisplayName type = function
Function details: 
lineNumber: 26
columnNumber: 40
scriptId is valid: true
functionName: user-friendly name

Running: testGetDetailsOfFunctionWithDisplayNameGetter
functionWithDisplayNameGetter type = function
Function details: 
lineNumber: 29
columnNumber: 46
scriptId is valid: true
functionName: functionWithDisplayNameGetter

Running: testSmallClosure
smallClosure type = function
Function details: 
lineNumber: 32
columnNumber: 57
scriptId is valid: true
functionName: 

Running: testBigClosure
bigClosure type = function
Function details: 
lineNumber: 45
columnNumber: 35
scriptId is valid: true
functionName: 

Running: testGenFunction
gen type = function
Function details: 
lineNumber: 52
columnNumber: 21
scriptId is valid: true
functionName: gen
isGenerator: true
